[PDK-5983] Fix review comments of IPC module

+++ packages/ti/drv/ipc/soc/V0/ipc_soc.c
@@ -253,7 +253,7 @@ int32_t Ipc_setCoreEventId(uint32_t selfId, Ipc_MbConfig* cfg, uint32_t intrCnt)
         vimEventBaseNum = (start + range) - offset;
 
         /* Translation must happen after this offset */
-        retVal = Ipc_sciclientIrqTranslate(selfId, vimEventBaseNum,
+        retVal = Ipc_sciclientIrqTranslate((uint16_t)selfId, vimEventBaseNum,
                                            &proc_irq);
         if (CSL_PASS == retVal)
         {
@@ -397,8 +397,8 @@ int32_t Ipc_sciclientIrqTranslate(uint16_t coreId, uint32_t eventId,
         uint16_t *procIrq)
 {
     return Sciclient_rmIrqTranslateIrOutput(req_type[coreId],
-                                            eventId,
-                                            map_dst_id[coreId],
+                                            (uint16_t)eventId,
+                                            (uint16_t)map_dst_id[coreId],
                                             procIrq);
 }
 
@@ -417,9 +417,9 @@ int32_t Ipc_sciclientIrqRelease(uint16_t coreId, uint32_t clusterId,
                               TISCI_MSG_VALUE_RM_DST_HOST_IRQ_VALID;
     rmIrqRel.src_id         = map_src_id[clusterId];
     rmIrqRel.src_index      = userId;
-    rmIrqRel.dst_id         = map_dst_id[coreId];
-    rmIrqRel.dst_host_irq   = intNumber;
-    rmIrqRel.secondary_host = TISCI_MSG_VALUE_RM_UNUSED_SECONDARY_HOST;
+    rmIrqRel.dst_id         = (uint16_t)map_dst_id[coreId];
+    rmIrqRel.dst_host_irq   = (uint16_t)intNumber;
+    rmIrqRel.secondary_host = (uint8_t)TISCI_MSG_VALUE_RM_UNUSED_SECONDARY_HOST;
 
     retVal = Sciclient_rmIrqRelease(
                  &rmIrqRel, IPC_SCICLIENT_TIMEOUT);
@@ -443,9 +443,9 @@ int32_t Ipc_sciclientIrqSet(uint16_t coreId, uint32_t clusterId,
     rmIrqReq.valid_params   = TISCI_MSG_VALUE_RM_DST_ID_VALID |
                               TISCI_MSG_VALUE_RM_DST_HOST_IRQ_VALID;
     rmIrqReq.src_id         = map_src_id[clusterId];
-    rmIrqReq.src_index      = userId;
-    rmIrqReq.dst_id         = map_dst_id[coreId];
-    rmIrqReq.dst_host_irq   = intNumber;
+    rmIrqReq.src_index      = (uint16_t)userId;
+    rmIrqReq.dst_id         = (uint16_t)map_dst_id[coreId];
+    rmIrqReq.dst_host_irq   = (uint16_t)intNumber;
     rmIrqReq.secondary_host = TISCI_MSG_VALUE_RM_UNUSED_SECONDARY_HOST;
 
     /* Config event */
@@ -463,7 +463,7 @@ int32_t Ipc_getIntNumRange(uint32_t coreIndex,
     struct tisci_msg_rm_get_resource_range_req  req = {0};
 
     req.type           = req_type[coreIndex];
-    req.subtype        = req_subtype[coreIndex];
+    req.subtype        = (uint8_t)req_subtype[coreIndex];
     req.secondary_host = TISCI_MSG_VALUE_RM_UNUSED_SECONDARY_HOST;
 
     res.range_num = 0;
@@ -477,7 +477,7 @@ int32_t Ipc_getIntNumRange(uint32_t coreIndex,
     if (CSL_PASS != retVal || res.range_num == 0) {
         /* Try with HOST_ID_ALL */
         req.type           = req_type[coreIndex];
-        req.subtype        = req_subtype[coreIndex];
+        req.subtype        = (uint8_t)req_subtype[coreIndex];
         req.secondary_host = TISCI_HOST_ID_ALL;
 
         retVal = Sciclient_rmGetResourceRange(
